exp/test-prefix: add 'command' to mpirun
This is for ANL Theta, where wrappe scripts need to tell
apart the arguments to mpirun from the program+args being invoked.

We choose 'command' over '--' in order to keep the the interface to the wrapper
mpirun the same as to the original mpirun ('command' is effectively a no-op in
Bash, in our context).

This should be harmless for other clusters.