Revert "ptools: anl-theta: wrapper for MPI compiler wrappers"
This reverts commit ab5d474dc0cfb4267bc9dd0631feec3279e7f34e.

This is unnecessary when marking the MCA components of OpenMPI that depend on
ALPS libs to be built in DSO mode instead of static.  I knew that, but the
issue was that I turned off the openmpi_fabrics_xpmem use flag for an
unrelated reason, which turned off the needed DSO setting for the sm component,
but due to OpenMPI bug, `--without-xpmem` is not honored, so we statically
linked the sm component which still ended up with dep on libxpmem. Bleh.