Revert "ptools: anl-theta: disable sm/xpmem transport"
This reverts commit 0eb8610621b14533b5025644a3910be1c7bba3c4.

Can't disable sm transport like this. This forces uGNI transport
to be required, which breaks when compiling using CASPER compiler
on worker nodes, where Cray libraries are not found due to different
version than on login nodes (where OpenMPI was built). When no
transport is forced, failure to load uGNI still happens but is
not fatal.

As a resolution to the segfault with xpmem, we fixed sys-cluster/openmpi ebuild
to completely disable xpmem at build time.