distfiles: don't pack the repos, will add directly
Packing is inefficient when it comes to updates. Updates to
git repos are very efficient in terms of space, since they
just add blobs in an append-only fashion. But if we pack repos
into tarballs, we lose that, and get a totally different
tarball after every update.

Also, it's just extra complexity. Let's trust LFS to be fast enough to
deal with many small files (it's only about 7000 extra files).