bootstrap.sh: leave LFS enabled for the repo
Even though the bootstrapped git-lfs will not be in PATH,
ultimately we do want it enabled. Otherwise, when it
will come time to git pull, the LFS objects won't get
fetched, which is worse than the lesser evil. The lesser
evil only arises if the user wants to do git operations
outside of Prefix (i.e. before the Prefix was built);
under normal circumstances, this shouldn't be necessary,
and under abnormal circumstances it is easy enough to
disable LFS manually as needed.